01_Lchain/pyproject.toml
[project]
name = "LChain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "huggingface-hub>=0.33.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.0.6",
    "ollama>=0.5.1",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "transformers>=4.52.4",
]

[tool.uv.sources]
LChain = { workspace = true }
